@@ -1,6 +1,6 @@
 require('../lib/menu');
 
-var React   = require('react/addons');
+var React   = require('react');
 var Reflux  = require('reflux');
 var _       = require('lodash');
 
